+++ Multi-Threading/MultiThreadingSample/Default_Commands.txt
@@ -1,59 +1,65 @@
-TotalThreads=100
-activeThreads=5 
-BaseInit=true 
-processes=[NonAPDFL, PDFa, PDFx, XPS2PDF, TextExtract, Rasterizer, Flattener]
-PauseEvery=0
 LogFile=MTRun.log
+LibraryPath=c:\Datalogics\APDFL15.0.4\CPlusPlus\Binaries
+PluginsPath=c:\Datalogics\APDFL15.0.4\CPlusPlus\Binaries
+ResourcePath=c:\Datalogics\APDFL15.0.4\Resources
+DataPath=LibraryPath=c:\Datalogics\APDFL15.0.4\Resources\Sample_input
+TotalThreads=100
+ActiveThreads=5
+PauseEvery=0
+BaseInit=true
+processes=[PDFa, PDFx, XPS2PDF, TextExtract, Rasterizer, Flattener]
 TempMemFileSys=false
+silent=true
 NonAPDFLOptions=[
-       silent=true, 
-       noapdfl=true, 
-       LoadPlugins=false,
        InFileName=%AddRedaction.pdf,
        OutFilePath=Output,
+       silent=true, 
+       LoadPlugins=false,
+       noapdfl=true,
        Repetitions=5,
        Primes=1000
                 ]
 PDFaOptions=[
-       silent=true, 
-       noapdfl=false, 
        InFileName=%trans_1page.pdf,
        OutFilePath=Output,
+       silent=true, 
+       LoadPlugins=true,
+       noapdfl=false, 
        ConvertOption=3,
        LoadPlugins=true,
        RasterizeFontErrors=true,
        RemoveAllAnnotations=false
                      ]
 PDFxOptions=[
-       silent=true, 
-       noapdfl=false, 
-       LoadPlugins=true,
        InFileName=%AddRedaction.pdf,
        OutFilePath=Output,
+       silent=true, 
+       LoadPlugins=true,
+       noapdfl=false, 
        ConvertOption=1,
        RemoveAllAnnotations=false
                    ]
 XPS2PDFOptions=[
-       silent=true, 
-       noapdfl=false, 
-       LoadPlugins=true,
        InFileName=%XPStoPDF.xps,
        OutFilePath=Output
+       silent=true, 
+       LoadPlugins=true,
+       noapdfl=false,
                       ]
 TextExtractOptions=[
-       silent=true, 
-       noapdfl=false, 
-       LoadPlugins=false,
        InFileName=%constitution.pdf,
        OutFilePath=Output,
+       silent=true, 
+       LoadPlugins=false,
+      noapdfl=false, 
        NumberOfPages=1
        ]
 RasterizerOptions=[
-       silent=true, 
-       noapdfl=false, 
-       LoadPlugins=false,
        InFileName=%constitution.pdf,
        OutFilePath=Output,
+       silent=true, 
+       LoadPlugins=false,
+       noapdfl=false,
        NumberOfPages=1,
        repetitions=1,
        SaveImages=false,
@@ -61,10 +67,10 @@ RasterizerOptions=[
        ColorModel=DeviceRGB
                     ]
 FlattenOptions=[
-       silent=true, 
-       noapdfl=false, 
-       LoadPlugins=true,
        InFileName=%trans_multipage.pdf,
        OutFilePath=Output,
+       silent=true, 
+       LoadPlugins=true,
+       noapdfl=false, 
        NumberOfPages=1
                     ]
