+++ Multi-Threading/MultiThreadingSample/MTHeader.h
@@ -4,6 +4,8 @@
 //
 //=====================================================================
 // This header file defines classs used in the multiThreading sample.
+#ifndef MTHEADER_H
+#define MTHEADER_H
 
 #include <iostream>
 #include <fstream>
@@ -21,6 +23,9 @@
 #include <sys/types.h>
 #include <sys/stat.h>
 #endif
+
+#include "ASExpT.h"
+#include "Utilities.h"
 
 
 /* These two classes are used to implement a command line parser, for the syntax:
@@ -506,449 +511,4 @@ typedef pthread_mutex_t *CSMutex;
 #endif
 
 
-/* Thread Communication */
-class workerclass;
-
-typedef struct
-{
-    ASInt32         threadNumber;                       /* Serial number of thread in set of threads */
-    ASUns32         sequence;
-    SDKThreadID     threadID;                           /* Platform dependent thread "handle" */
-    void           *object;                             /* Worker Thread Object */
-    int             objectType;                         /* Enumerator giving thread type */
-    APDFLib        *instance;                           /* APDFL Library instance */
-    ASInt32         result;                             /* Numeric result, unique to worker type. But "zero" is always "No Problem" */
-    time_t          startTime, endTime;                 /* Used in Unix only, wall time started/stopped */
-    clock_t         startCPU, endCPU;                   /* Used in Unix only to track CPU time used. */
-    double          wallTimeUsed, cpuTimeUsed;          /* Walltime start to finish, and CPU time (user and kernal) consumed */
-    double          percentUtilized;                    /* Percentage of CPU time in wall time */
-    bool            silent;                             /* When true, write nothing to stdout! */
-    bool            noAPDFL;                            /* When true, do not init/term the library in this thread! */
-    bool            threadCompleted;                    /* Mark the thread complete for unix to locate it */
-    FILE           *logFile;                            /* Write status message to this file */
-    bool            logFileSet;                         /* If log file is set, then default "silent" to "false". */
-    bool            LoadPlugins;                        /* If true, we must load plugins for this type of worker. */
-    bool            UseTempMemFileSys;                  /* If true, use the Ram File Sys for temp files. */
-} ThreadInfo;
-
-
-/* This is a base class for the worker threads
-*/
-class workerclass
-{
-public:
-    /* Every worker thread needs an input and output file 
-    ** The ideal worker has a sequence of them, so that we may 
-    ** eiher select all workers to be identical, or force them to behave differently, 
-    ** and reveal diferent conflicts.
-    ** These will be set in the standard options logic (workerclass::ParserOptions()) from the command
-    ** line keyword "InFileName".
-    */
-    int          InFileCount;
-    char       **InFilePath;
-    char       **InFileName;
-    char       **InFileSuffix;
-
-    /* Some classes (Ergo: merge documents) may have more than one input file
-    ** The following two options are established to allow for 2 or three input
-    ** files
-    */
-    int          InFile2Count;
-    char       **InFile2Path;
-    char       **InFile2Name;
-    char       **InFile2Suffix;
-    int          InFile3Count;
-    char       **InFile3Path;
-    char       **InFile3Name;
-    char       **InFile3Suffix;
-
-    /* I don't ever see a real reason for seperate output paths,  But leave this for Symmetry 
-    ** These will be set in the standard options logic (workerclass::ParserOptions()) from the command
-    ** line keyword "OutFilePath".
-    */
-    int          OutPathCount;
-    char       **OutFilePath;
-
-    /* When true (default) do not print messages from this object 
-    ** This will be set in the standard options logic (workerclass::ParserOptions()) from the command
-    ** line keyword "Silent".*/
-    bool        silent;
-
-    /* When true (Default false) do not initiate/terminate the library per thread!
-    ** (This may be set to true by default for a given worker class, overridding both 
-    **  the setting from the options, and the default) 
-    ** This will be set in the standard options logic (workerclass::ParserOptions()) from the command
-    ** line keyword "noAPDFL".
-    */
-    bool        noAPDFL;
-
-    /* Dictionary of options for this object */
-    attributes *threadAttributes;
-
-    workerclass () { InFileCount = 0;
-                     InFile2Count = 0;
-                     InFile3Count = 0;
-                     OutPathCount = 0;
-                     silent = true;
-                     noAPDFL = false;
-                     InFilePath = InFileName = InFileSuffix = OutFilePath = NULL;
-    }
-    ~workerclass () { for (int index = 0; index < InFileCount; index++) 
-                          { free (InFilePath[index]); free (InFileName[index]); free (InFileSuffix[index]); }
-                      for (int index = 0; index < InFile2Count; index++)
-                          { free (InFile2Path[index]); free (InFile2Name[index]); free (InFile2Suffix[index]); }
-                      for (int index = 0; index < InFile3Count; index++)
-                          {free (InFile3Path[index]); free (InFile3Name[index]); free (InFile3Suffix[index]); }
-                      if (InFileCount) free (InFilePath); 
-                      if (InFileCount) free (InFileName); 
-                      if (InFileCount) free (InFileSuffix); 
-                      if (InFile2Count) free (InFile2Path);
-                      if (InFile2Count) free (InFile2Name);
-                      if (InFile2Count) free (InFile2Suffix);
-                      if (InFile3Count) free (InFile3Path);
-                      if (InFile3Count) free (InFile3Name);
-                      if (InFile3Count) free (InFile3Suffix);
-                      for (int index = 0; index < OutPathCount; index++)
-                          { free (OutFilePath[index]);}
-                      if (OutPathCount) free (OutFilePath);
-                      delete threadAttributes;
-    }
-
-    void WorkerThread (ThreadInfo *)
-    {
-        return;
-    }
-
-    char *GetInFileName (int threadSequence)
-    {
-        char workname[2048], *result;
-        int index = threadSequence % InFileCount;
-        sprintf (workname, "%s%c%s.%s", InFilePath[index], PathSep, InFileName[index], InFileSuffix[index]);
-        result = (char *)malloc (strlen (workname) + 1);
-        strcpy (result, workname);
-        return (result);
-    }
-
-    char *GetInFile2Name (int threadSequence)
-    {
-        if (InFile2Count == 0)
-            return (NULL);
-
-        char workname[2048], *result;
-        int index = threadSequence % InFile2Count;
-        sprintf (workname, "%s%c%s.%s", InFile2Path[index], PathSep, InFile2Name[index], InFile2Suffix[index]);
-        result = (char *)malloc (strlen (workname) + 1);
-        strcpy (result, workname);
-        return (result);
-    }
-
-
-    char *GetInFile3Name (int threadSequence)
-    {
-        if (InFile3Count == 0)
-            return (NULL);
-
-        char workname[2048], *result;
-        int index = threadSequence % InFile3Count;
-        sprintf (workname, "%s%c%s.%s", InFile3Path[index], PathSep, InFile3Name[index], InFile3Suffix[index]);
-        result = (char *)malloc (strlen (workname) + 1);
-        strcpy (result, workname);
-        return (result);
-    }
-
-
-    char * GetOutFileName (int threadSequence, int inner = -1)
-    {
-        int indexOut = threadSequence % OutPathCount;
-        int indexIn = threadSequence % InFileCount;
-        char workname[2048], *result;;
-        if (inner == -1)
-            sprintf (workname, "%s%c%s_%01d.%s", OutFilePath[indexOut], PathSep, InFileName[indexIn], threadSequence + 1, InFileSuffix[indexIn]);
-        else
-            sprintf (workname, "%s%c%s_%01d_%01d.%s", OutFilePath[indexOut], PathSep, InFileName[indexIn], threadSequence + 1, inner + 1, InFileSuffix[indexIn]);
-        result = (char *)malloc (strlen (workname) + 1);
-        strcpy (result, workname);
-        return (result);
-    }
-
-    void startThreadWorker (ThreadInfo *info)
-    {
-#ifndef WIN_PLATFORM
-        info->startTime = time (&info->startTime);
-        info->startCPU = clock ();
 #endif
-        if (noAPDFL)
-        {
-            info->instance = NULL;
-            info->noAPDFL = true;
-        }
-        else
-        {
-            ASUns32 flags = 0;
-            if (!info->LoadPlugins)
-                flags |= kDontLoadPlugIns;
-            info->instance = new APDFLib (flags);
-            if (info->UseTempMemFileSys)
-                ASSetTempFileSys (ASGetRamFileSys ());
-        }
-        info->silent = silent;
-    }
-
-    void endThreadWorker (ThreadInfo *info)
-    {
-        if (info->instance)
-            delete info->instance;
-
-#ifndef WIN_PLATFORM
-        info->endTime= time (&info->endTime);
-        info->endCPU = clock ();
-        info->wallTimeUsed = ((info->endTime - info->startTime) * 1.0) / CLOCKS_PER_SEC;
-        info->cpuTimeUsed = ((info->endCPU - info->startCPU) * 1.0) / CLOCKS_PER_SEC;
-        // yluo doneThread is undefined here
-        // doneThread->percentUtilized = (doneThread->cpuTimeUsed / dontThread->wallTimeUsed) * 100;
-#endif
-
-        info->threadCompleted = true;
-    }
-
-
-    void splitpath (char *path, char **toPath, char **filename, char **suffix)
-    {
-        static char pathCopy[2048];
-        strcpy (pathCopy, path);
-        char *finder = pathCopy + strlen (pathCopy) - 1;
-        *suffix = *filename = *toPath = NULL;
-        while (finder != pathCopy)
-        {
-            while ((finder > pathCopy) && (finder[0] != '.') & (finder[0] != PathSep))
-                finder--;
-            if (finder[0] == '.')
-            {
-                *suffix = finder + 1;
-                finder[0] = 0;
-                finder--;
-                continue;
-            }
-            if (finder[0] == PathSep)
-            {
-                *filename = finder + 1;
-                finder[0] = 0;
-                finder--;
-                *toPath = pathCopy;
-            }
-            break;
-        }
-    }
-
-    /* If we get here, and the file name start with a 
-    ** percent sign, then expand the name by preceedingit with the
-    ** standard path to sample input
-    */
-    char *expandFileName (char *inname)
-    {
-        char *result;
-        if (inname[0] == '%')
-        {
-            result = (char *)malloc (strlen (inname) + strlen (inputPath) + 1);
-            sprintf (result, "%s%c%s", inputPath, PathSep, &inname[1]);
-        }
-        else
-        {
-            result = (char *)malloc (strlen (inname) + 1);
-            strcpy (result, inname);
-        }
-        return (result);
-    }
-
-    void parseOneFileName (int index, char *inname)
-    { 
-        char *expandedName = expandFileName (inname);
-
-        char *path, *name, *suffix;
-        workerclass::splitpath (expandedName, &path, &name, &suffix);
-
-        free (expandedName);
-
-        InFilePath[index] = (char *)malloc (strlen (path) + 1);
-        strcpy (InFilePath[index], path);
-        InFileName[index] = (char *)malloc (strlen (name) + 1);
-        strcpy (InFileName[index], name);
-        InFileSuffix[index] = (char *)malloc (strlen (suffix) + 1);
-        strcpy (InFileSuffix[index], suffix);
-    }
-
-    void parseTwoFileName (int index, char *inname)
-    {
-        char *expandedName = expandFileName (inname);
-
-        char *path, *name, *suffix;
-        workerclass::splitpath (expandedName, &path, &name, &suffix);
-
-        free (expandedName);
-
-        InFile2Path[index] = (char *)malloc (strlen (path) + 1);
-        strcpy (InFilePath[index], path);
-        InFile2Name[index] = (char *)malloc (strlen (name) + 1);
-        strcpy (InFileName[index], name);
-        InFile2Suffix[index] = (char *)malloc (strlen (suffix) + 1);
-        strcpy (InFile2Suffix[index], suffix);
-    }
-
-    void parseThreeFileName (int index, char *inname)
-    {
-        char *expandedName = expandFileName (inname);
-
-        char *path, *name, *suffix;
-        workerclass::splitpath (expandedName, &path, &name, &suffix);
-
-        free (expandedName);
-
-        InFile3Path[index] = (char *)malloc (strlen (path) + 1);
-        strcpy (InFilePath[index], path);
-        InFile3Name[index] = (char *)malloc (strlen (name) + 1);
-        strcpy (InFile3Name[index], name);
-        InFile3Suffix[index] = (char *)malloc (strlen (suffix) + 1);
-        strcpy (InFile3Suffix[index], suffix);
-    }
-
-    void    ParseOptions (valuelist *values, char *defaultInFileName, char *defaultOutFilePath)
-    {
-
-        if (values)
-        {
-            /* Parse the attributes as is there were a command line */
-            int listSize = values->size ();
-            char **valueArray = (char **)malloc ((listSize + 1)* sizeof (char *));
-            valueArray[0] = NULL;
-            for (int count = 0; count < values->size (); count++)
-                valueArray[count + 1] = values->value (count);
-            threadAttributes = new attributes (listSize + 1, valueArray);
-        }
-        else
-        {
-            char *nullPtr = NULL;
-            threadAttributes = new attributes (1, &nullPtr);
-        }
-
-        if (threadAttributes->IsKeyPresent ("InFileName"))
-        {
-            valuelist *nameValues = threadAttributes->GetKeyValue ("InFileName");
-            InFileCount = nameValues->size ();
-            InFilePath = (char **)malloc (sizeof (char*) * InFileCount);
-            InFileName = (char **)malloc (sizeof (char*) * InFileCount);
-            InFileSuffix = (char **)malloc (sizeof (char*) * InFileCount);
-
-            for (int index = 0; index < InFileCount; index++)
-                parseOneFileName (index, nameValues->value (index));
-        }
-        else
-        {
-            InFileCount = 1;
-            InFilePath = (char **)malloc (sizeof (char*));
-            InFileName = (char **)malloc (sizeof (char*));
-            InFileSuffix = (char **)malloc (sizeof (char*));
-            parseOneFileName (0, defaultInFileName);
-        }
-
-        if (threadAttributes->IsKeyPresent ("InFile2Name"))
-        {
-            valuelist *nameValues = threadAttributes->GetKeyValue ("InFile2Name");
-            InFile2Count = nameValues->size ();
-            InFile2Path = (char **)malloc (sizeof (char*) * InFile2Count);
-            InFile2Name = (char **)malloc (sizeof (char*) * InFile2Count);
-            InFile2Suffix = (char **)malloc (sizeof (char*) * InFile2Count);
-
-            for (int index = 0; index < InFileCount; index++)
-                parseTwoFileName (index, nameValues->value (index));
-        }
-
-        if (threadAttributes->IsKeyPresent ("InFile3Name"))
-        {
-            valuelist *nameValues = threadAttributes->GetKeyValue ("InFile3Name");
-            InFile3Count = nameValues->size ();
-            InFile3Path = (char **)malloc (sizeof (char*) * InFile3Count);
-            InFile3Name = (char **)malloc (sizeof (char*) * InFile3Count);
-            InFile3Suffix = (char **)malloc (sizeof (char*) * InFile3Count);
-
-            for (int index = 0; index < InFileCount; index++)
-                parseThreeFileName (index, nameValues->value (index));
-        }
-
-        if (threadAttributes->IsKeyPresent ("OutFilePath"))
-        {
-            valuelist *nameValues = threadAttributes->GetKeyValue ("OutFilePath");
-            OutPathCount = nameValues->size ();
-            OutFilePath = (char **)malloc (sizeof (char *) * OutPathCount);
-            for (int index = 0; index < OutPathCount; index++)
-            {
-                char *value = nameValues->value (index);
-                OutFilePath[index] = (char *)malloc (strlen (value) + 1);
-                strcpy (OutFilePath[index], value);
-            }
-        }
-        else
-        {
-            OutFilePath = (char **)malloc (sizeof (char *));
-            OutPathCount = 1;
-            OutFilePath[0] = (char *)malloc (strlen (defaultOutFilePath) + 1);
-            strcpy (OutFilePath[0], defaultOutFilePath);
-        }
-
-        if (threadAttributes->IsKeyPresent ("Silent"))
-            silent = threadAttributes->GetKeyValueBool ("Silent");
-
-        if (threadAttributes->IsKeyPresent ("NoAPDFL"))
-            noAPDFL = threadAttributes->GetKeyValueBool ("NoAPDFL");
-
-        char workpath[2048];
-        for (int index = 0; index < InFileCount; index++)
-        {
-            sprintf (workpath, "%s%c%s.%s", InFilePath[index], PathSep, InFileName[index], InFileSuffix[index]);
-            if (access (workpath, 04))
-            {
-                printf ("The input file does not exist? \n%   \"%s\"\n", workpath);
-                exit (-1);
-            }
-        }
-
-        for (int index = 0; index < InFile2Count; index++)
-        {
-            sprintf (workpath, "%s%c%s.%s", InFile2Path[index], PathSep, InFile2Name[index], InFile2Suffix[index]);
-            if (access (workpath, 04))
-            {
-                printf ("The input file does not exist? \n%   \"%s\"\n", workpath);
-                exit (-1);
-            }
-        }
-
-        for (int index = 0; index < InFile3Count; index++)
-        {
-            sprintf (workpath, "%s%c%s.%s", InFile3Path[index], PathSep, InFile3Name[index], InFile3Suffix[index]);
-            if (access (workpath, 04))
-            {
-                printf ("The input file does not exist? \n%   \"%s\"\n", workpath);
-                exit (-1);
-            }
-        }
-
-        for (int index = 0; index < OutPathCount; index++)
-        {
-            if (access (OutFilePath[index], 6))
-            {
-#ifdef WIN_PLATFORM
-                mkdir (OutFilePath[index]);
-#else
-                mkdir (OutFilePath[index], 0777);
-#endif
-                if (access (OutFilePath[index], 6))
-                {
-                    printf ("The output path cannot be found or created!\n    \"%s\"\n", OutFilePath);
-                    exit (-1);
-                }
-            }
-        }
-    }
-
-
-};
-
