+++ Multi-Threading/MultiThreadingSample/MultiThreadingSample.cpp
@@ -114,23 +114,23 @@
 */
 
 /* This is the current set of known workers 
-** To this list, add the class name here,
+** To this list, define the new class in a new .cpp/.h file
+**               Add the class name to the list of includes here
 **               Add a field to the union WorkerClassPtr
-**               Add a value to the enumertor (These must be a solid set of numbers)
-**               Add an entry to the Workers table, giving a textual name to the class,
-**                  and a name for it's command line options, 
-**               Add a new class derived from workerclass. This class must have a ParseOptions method, 
-**                  and a WorkerThread method.
-**               Create an initialize an instance of the worker class inthe main procedure.
-**               Add a class to the switch in OuterWorker()
+**               Add a value to the enumertor in Worker.h (These must be a solid set of numbers,
+**                  The last number must be "NumberOfWorkers".)
+**               Add a class to the switch in OuterWorker()**
+**               Create and initialize an instance of the worker class in the main procedure.
+
 */
-class NonAPDFLWorker;
-class PDFaWorker ;
-class PDFxWorker;
-class XPS2PDFWorker;
-class TextextWorker;
-class RasterizerWorker;
-class FlattenWorker;
+#include "Worker.h"                 /* The base worker class */
+#include "NonAPDFL_Worker.h"
+#include "PDFA_Worker.h"
+#include "PDFX_Worker.h"
+#include "XPS2PDF_Worker.h"
+#include "TextExtract_Worker.h"
+#include "Rasterizer_Worker.h"
+#include "Flattener_Worker.h"
 
 typedef union workerclassptr
 {
@@ -144,1359 +144,13 @@ typedef union workerclassptr
 } WorkerClassPtr;
 
 
-typedef enum
-{
-    NONAPDFL,
-    PDFA,
-    PDFX,
-    XPS2PDF,
-    TextExtract,
-    Rasterizer,
-    Flattener
-};
 
-typedef struct worktypes
-{
-    char                *name;
-    char                *paramName;
-    int                  sequence;
-    bool                 LoadPlugins;
-} WorkerType;
-
-
-WorkerType workers[] = {
-
-/*      Name             Options             Id         Load Plugins*/
-    {"NonAPDFL",    "NonAPDFLOptions",       NONAPDFL,      false},
-    {"PDFa",        "PDFaOptions",           PDFA,          true},
-    {"PDFx",        "PDFxOptions",           PDFX,          true},
-    {"XPS2PDF",     "XPS2PDFOptions",        XPS2PDF,       true},
-    {"TextExtract", "TextExtractOptions",    TextExtract,   false},
-    {"Rasterizer",  "RasterizerOptions",     Rasterizer,    false},
-	{"Flattener",   "FlattenOptions",        Flattener,      true}
-};
-#define NumberOfWorkers (sizeof (workers)/sizeof (WorkerType))
-
+WorkerType workers[NumberOfWorkers];
 WorkerClassPtr workerClasses[NumberOfWorkers];
 
 
-/* These are the worker class objects that define the process to be 
-** executed in a given thread
-**
-** The set can be expanded indefinately.
-*/
-
-
-/*
-** NonAPDFL Worker
-** This worker type will simply perform some I/O and some CPU intensive work. It is intended to validate the 
-** framework itself, and to provide a baseline against other workers. 
-**
-** If "NoAPDFL=true", the thread will not even init/term the library. If false, it will Init/Term the library, and give us a picture of behaviour
-** doing only init term. Setting "NoAPDFL=true, Reptitions=0" will do nothing EXCEPT init/term the libraries!
-**
-**  NonAPDFLOptions=[
-**                   silent=true                                        When true, do not display status lines
-**                   InFileName=[test data path\AddRedactions.pdf]      A list of file to copy, One file will be copied by each thread, sequcence % #files.
-**                   OutFilePath=[Output]                               Directory where output is written
-**                   noAPDFL=false                                      When true, we will NOT init/term the library for each thread
-**                   Repetitions=[5]                                    How many time shall we do the inner loop (100 values max!)
-**                   Primes=[1000]                                      How many prime numbers shall we find in CPU loading (100 values max!)
-*/
-class NonAPDFLWorker : public workerclass
-{
-    /* This thread will copy a specified file N times, into new specified files.
-    ** Each time the file is copied, it will also find the fist 1000 prime numbers!
-    */
-public:
-    NonAPDFLWorker () { Repetitions[0] = 0; RepetitionsCount = 0; Primes[0] = 0; PrimesCount = 0; };
-    ~NonAPDFLWorker () { };
-
-    /* Parse the non-APDFL conversion thread options into attributes */
-    void ParseOptions (valuelist *values)
-    {
-        workerclass::ParseOptions (values, "%AddRedaction.pdf", "Output");
-
-        if (threadAttributes->IsKeyPresent ("LoadPlugins"))
-            workers[NONAPDFL].LoadPlugins = threadAttributes->GetKeyValueBool ("LoadPlugins");
-
-        if (threadAttributes->IsKeyPresent ("Repetitions"))
-        {
-            valuelist *values = threadAttributes->GetKeyValue ("Repetitions");
-            RepetitionsCount = values->size ();
-            for (int index = 0; index < RepetitionsCount; index++)
-                Repetitions[index] = values->GetValueInt (index);
-        }
-        else
-        {
-            RepetitionsCount = 1;
-            Repetitions[0] = 5;
-        }
-
-        if (threadAttributes->IsKeyPresent ("Primes"))
-        {
-            valuelist *values = threadAttributes->GetKeyValue ("Primes");
-            PrimesCount = values->size ();
-            for (int index = 0; index < PrimesCount; index++)
-                Primes[index] = values->GetValueInt (index);
-        }
-        else
-        {
-            PrimesCount = 1;
-            Primes[0] = 1000;
-        }
-
-    };
-
-    int FindPrimes (ASUns32 *primes, ASUns32 limit)
-    {
-        ASUns32 primeCount = 3;
-        primes[0] = 1;
-        primes[1] = 2;
-        primes[2] = 3;
-        for (ASUns32 isItPrime = 4; primeCount < limit; isItPrime++)
-        {
-            ASBool notPrime = false;
-            for (ASUns32 lastPrime = 1; lastPrime < primeCount; lastPrime++)
-            {
-                if (isItPrime % primes[lastPrime] == 0)
-                {
-                    notPrime = true;
-                    break;
-                }
-            }
-            if (notPrime)
-                continue;
-            primes[primeCount] = isItPrime;
-            primeCount++;
-        }
-        return (primeCount);
-    }
-
-
-    /* One thread worker procedure */
-    void WorkerThread (ThreadInfo *info)
-    {
-        int sequence = info->sequence;
-        if (!silent)
-            fprintf (info->logFile, "Non APDFL Worker Thread started! (Sequence: %01d, Thread: %01d\n", sequence+1, info->threadNumber + 1);
-
-        /* Presume we willcomplete cleanly */
-        info->result = 0;
-
-        /* This process will open an input file, read it's contents to memory, close the file, and write it N times into
-        ** new files in the output directory
-        */
-        char *fullFileName = GetInFileName (sequence);
-
-        /* Open the file, and get it's size */
-        FILE *input = fopen (fullFileName, "rb");
-
-        /* free the file name */
-        free (fullFileName);
-
-        if (!input)
-            /* if we could not open the file, mark as failed or reason 1 */
-            info->result = 1;
-        else
-        {
-            fseek (input, 0, SEEK_END);
-            size_t fileSize = ftell (input);
-            fseek (input, 0, SEEK_SET);
-
-            /* Read the files content into a memory buffer */
-            char *buffer = (char *)malloc (fileSize);
-            size_t bytesRead = fread (buffer, 1, fileSize, input);
-
-            /* Close the file */
-            fclose (input);
-
-            if (bytesRead != fileSize)
-            {
-                /* If we could not read the entire file, mark as failed for reason 2*/
-                info->result = 2;
-                free (buffer);
-            }
-            else
-            {
-                /* Create n new copies of the file */
-                for (int x = 0; x < Repetitions[sequence % RepetitionsCount]; x++)
-                {
-                    /* Build file name from options */
-                    fullFileName = GetOutFileName (sequence, x);
-
-                    /* Open the output file */
-                    FILE *output = fopen (fullFileName, "wb");
-
-                    /* Free the file name */
-                    free (fullFileName);
-
-                    if (!output)
-                        /* If we could not open the output, fail for reason 3 */
-                        info->result = 3;
-                    else
-                    {
-                        size_t writeBytes = fwrite (buffer, 1, fileSize, output);
-                        if (writeBytes != fileSize)
-                            /* if we could not write the entire file, fail for reason 4 */
-                            info->result = 4;
-                        fclose (output);
-                    }
-
-                    /* Burn some CPU as well */
-                    ASUns32 *primes = (ASUns32 *)malloc (sizeof (ASUns32) * Primes[sequence % PrimesCount]);
-                    ASUns32 primesFound = FindPrimes (primes, Primes[sequence % PrimesCount]);
-                }
-            }
-            free (buffer);
-        }
-
-        if (!silent)
-            fprintf (info->logFile, "Non APDFL Worker Thread completed! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-    }
-
-private:
-    ASInt32     Repetitions[100];
-    ASInt32     RepetitionsCount;
-    ASInt32     Primes[100];
-    ASInt32     PrimesCount;
-};
-
-/*
-** PDF/a worker
-** This thread willconvertone document to PDF/a.
-** NoAPDFL MUST be false, or the thread will not function.
-** InFileName as a list will use the Nth entry of the list, modulus the number of entries in the list, for each worker thread.
-
-**  PDFaOptions=[
-**                   silent=true                                        When true, do not display status lines
-**                   InFileName=[test data path\AddRedactions.pdf]      A list of file to copy, One file will be copied by each thread, sequcence % #files.
-**                   OutFilePath=[Output]                               Directory where output is written
-**                   noAPDFL=false                                      When true, we will NOT init/term the library for each thread
-**                   ConvertOption=[1]                                  Convertor Option selector, 1 to 4. (100 values max!)
-**                   RasterizeFontErrors=[false]                        SettingforPDF/a conversion option "rasterizeFontErrors" (100 values max!)
-**                   RemoveAllAnnotations=[false]                        SettingforPDF/a conversion option "removeAllAnnotations" (100 values max!)
-*/
-#include "PDFProcessorCalls.h"
-ASBool PDFProcessorProgressMonitorCB(ASInt32 pageNum, ASInt32 totalPages, float current, void *clientData)
-{
-    if (clientData)
-    {
-        ASBool * IsMonitorCalled = (ASBool*)clientData;
-        if (*IsMonitorCalled == false)
-        {
-            printf("PDFProcessor Progress Monitor CallBack\n");
-            //Set to true to Display this Message Only Once
-            *IsMonitorCalled = true;
-        }
-    }
-
-    printf("PDFProcessor Page %d of %d. Overall Progress = %f %%. \n",
-        pageNum + 1, /* Adding 1, since Page numbers are 0-indexed*/
-        totalPages,
-        current /* Current Overall Progress */);
-
-    //Return 1 to Cancel conversion
-    return 0;
-}
-class PDFaWorker : public workerclass
-{
-#define NumberOfPDFAConvertOptions 4
-
-public:
-    PDFaWorker () {
-                    ConvertOptions[0] = kPDFProcessorConvertToPDFA1aRGB;
-                    ConvertOptions[1] = kPDFProcessorConvertToPDFA1aCMYK;
-                    ConvertOptions[2] = kPDFProcessorConvertToPDFA1bRGB;
-                    ConvertOptions[3] = kPDFProcessorConvertToPDFA1bCMYK;
-                  };
-
-    ~PDFaWorker ()
-    { };
-
-    /* Parse the PDF/a conversion thread options into attributes */
-    void ParseOptions (valuelist *values)
-    {
-        workerclass::ParseOptions (values, "%AddRedaction.pdf", "Output");
-
-        if (threadAttributes->IsKeyPresent ("LoadPlugins"))
-            workers[PDFA].LoadPlugins = threadAttributes->GetKeyValueBool ("LoadPlugins");
-
-        if (threadAttributes->IsKeyPresent ("RasterizeFontErrors"))
-        {
-            valuelist *values = threadAttributes->GetKeyValue ("RasterizeFontErrors");
-            rasterizeFontErrorsCount = values->size ();
-            for (int index = 0; index < rasterizeFontErrorsCount; index++)
-                rasterizeFontErrors[index] = values->GetValueBool (index);
-        }
-        else
-        {
-            rasterizeFontErrorsCount = 1;
-            rasterizeFontErrors[0] = false;
-        }
-
-        if (threadAttributes->IsKeyPresent ("RemoveAllAnnotations"))
-        {
-            valuelist *values = threadAttributes->GetKeyValue ("RemoveAllAnnotations");
-            removeAllAnnotationsCount = values->size ();
-            for (int index = 0; index < removeAllAnnotationsCount; index++)
-                removeAllAnnotations[index] = values->GetValueBool (index);
-        }
-        else
-        {
-            removeAllAnnotationsCount = 1;
-            removeAllAnnotations[0] = false;
-        }
-
-        if (threadAttributes->IsKeyPresent ("ConvertOption"))
-        {
-            valuelist *values = threadAttributes->GetKeyValue ("ConvertOption");
-            convertorOptionsCount = values->size ();
-            for (int index = 0; index < convertorOptionsCount; index++)
-            {
-                /* The value should be specified as 1 to 4 
-                ** but must be here converted to 0 to 3
-                */
-                convertorOptions[index] = values->GetValueInt (index);
-                convertorOptions[index] -= 1;
-                convertorOptions[index] %= 4;
-            }
-        }
-        else
-        {
-            convertorOptionsCount = 1;
-            convertorOptions[0] = 0;
-        }
-
-    };
-
-    /* One thread worker procedure */
-    void WorkerThread (ThreadInfo *info)
-    {
-        int sequence = info->sequence;
-        if (!silent)
-            fprintf (info->logFile, "PDF/a Worker Thread Started! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-
-        /* Generate input and output file names */
-        char *fullFileName = GetInFileName (sequence);
-        char *fullOutputFileName = GetOutFileName (sequence, -1);
-
-        DURING
-            /* Open the input document */
-            APDFLDoc inDoc (fullFileName, true);
-
-            /* Free the input file names */
-            free (fullFileName);
-
-            /* Initialize the HFT for the plugin */
-            gPDFProcessorHFT = InitPDFProcessorHFT;
-
-            //initialize PDFProcessor plugin
-            if (!PDFProcessorInitialize ())
-                /* If the plugin cannot be initilized! */
-                info->result = 1;
-            else
-            {
-                /* COnstruct the user params recor for the PDF/A conversion */
-                PDFProcessorPDFAConvertParamsRec userParams;
-
-                memset ((char *)&userParams, 0, sizeof (PDFProcessorPDFAConvertParamsRec));
-                userParams.size = sizeof (PDFProcessorPDFAConvertParamsRec);
-                userParams.progMon = PDFProcessorProgressMonitorCB;
-                userParams.colorCompression = kPDFProcessorColorJpegCompression;
-                userParams.noRasterizationOnFontErrors = false; 
-                userParams.removeAllAnnotations = false;
-
-
-                /* Create the ouput file ASPath name */
-#if !MAC_ENV	
-                ASPathName destFilePath = ASFileSysCreatePathName (NULL, ASAtomFromString ("Cstring"), fullOutputFileName, NULL);
-#else
-                ASPathName destFilePath = APDFLDoc::makePath (fullOutputFileName);
-#endif
-
-                /* Release the output file name */
-                free (fullOutputFileName);
-                
-                /* Perform the conversions */
-                PDFProcessorConvertAndSaveToPDFA (inDoc.getPDDoc (), destFilePath, ASGetDefaultFileSys (), 
-                                ConvertOptions[convertorOptions[sequence % convertorOptionsCount]+1], &userParams);
-
-                /* Release the output path name */
-                ASFileSysReleasePath (NULL, destFilePath);
-
-                /* terminate the plugin */
-                PDFProcessorTerminate ();
-            }
-        HANDLER
-                info->result = 2;
-        END_HANDLER
-
-        if (!silent)
-            fprintf (info->logFile, "PDF/a Worker Thread Completed! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-    }
-
-private:
-    bool rasterizeFontErrors[100];
-    int  rasterizeFontErrorsCount;
-    bool removeAllAnnotations[100];
-    int  removeAllAnnotationsCount;
-    ASUns32 convertorOptions[100];
-    int  convertorOptionsCount;
-
-    PDFProcessorPDFAConversionOption ConvertOptions[4];
-};
-
-
-/*
-** PDF/x worker
-** This thread willconvert one document to PDF/x.
-** NoAPDFL MUST be false, or the thread will not function.
-** InFileName as a list will use the Nth entry of the list, modulus the number of entries in the list, for each worker thread.
-
-**  PDFxOptions=[
-**                   silent=true                                        When true, do not display status lines
-**                   InFileName=[test data path\AddRedactions.pdf]      A list of file to copy, One file will be copied by each thread, sequcence % #files.
-**                   OutFilePath=[Output]                               Directory where output is written
-**                   noAPDFL=false                                      When true, we will NOT init/term the library for each thread
-**                   ConvertOption=[1]                                  Convertor Option selector, 1 to 2. (100 values max!)
-**                   RemoveAllAnnotations=[false]                        SettingforPDF/a conversion option "removeAllAnnotations" (100 values max!)
-*/
-class PDFxWorker : public workerclass
-{
-public:
-    PDFxWorker () {
-                        ConvertOptions[0] = kPDFProcessorConvertToPDFX1a2001;
-                        ConvertOptions[1] = kPDFProcessorConvertToPDFX32003;
-                  };
-    ~PDFxWorker () { };
-
-    /* Parse the PDF/x conversion thread options into attributes */
-    void ParseOptions (valuelist *values)
-    {
-        workerclass::ParseOptions (values, "%AddRedaction.pdf", "Output");
-
-        if (threadAttributes->IsKeyPresent ("LoadPlugins"))
-            workers[PDFX].LoadPlugins = threadAttributes->GetKeyValueBool ("LoadPlugins");
-
-        if (threadAttributes->IsKeyPresent ("RemoveAllAnnotations"))
-        {
-            valuelist *values = threadAttributes->GetKeyValue ("RemoveAllAnnotations");
-            removeAllAnnotationsCount = values->size ();
-            for (int index = 0; index < removeAllAnnotationsCount; index++)
-                removeAllAnnotations[index] = values->GetValueBool (index);
-        }
-        else
-        {
-            removeAllAnnotationsCount = 1;
-            removeAllAnnotations[0] = false;
-        }
-
-        if (threadAttributes->IsKeyPresent ("ConvertorOption"))
-        {
-            valuelist *values = threadAttributes->GetKeyValue ("ConvertorOption");
-            convertorOptionsCount = values->size ();
-            for (int index = 0; index < convertorOptionsCount; index++)
-            {
-                /* The value should be specified as 1 to 2
-                ** but must be here converted to 0 to 1
-                */
-                convertorOptions[index] = values->GetValueInt (index);
-                convertorOptions[index] -= 1;
-                convertorOptions[index] %= 2;
-            }
-        }
-        else
-        {
-            convertorOptionsCount = 1;
-            convertorOptions[0] = 0;
-        }
-    };
-
-
-    /* One thread worker procedure */
-    void WorkerThread (ThreadInfo *info)
-    {
-        int sequence = info->sequence;
-
-        if (!silent)
-            fprintf (info->logFile, "PDF/x Worker Thread started! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-
-        /* Generate input and output file names */
-        char *fullFileName = GetInFileName (sequence);
-        char *fullOutputFileName = GetOutFileName (sequence, -1);
-
-        DURING
-            /* Open the input document */
-            APDFLDoc inDoc (fullFileName, true);
-
-            /* Free the input file names */
-            free (fullFileName);
-
-            /* Initialize the HFT for the plugin */
-            gPDFProcessorHFT = InitPDFProcessorHFT;
-
-            //initialize PDFProcessor plugin
-            if (!PDFProcessorInitialize ())
-                /* If the plugin cannot be initilized! */
-                info->result = 1;
-            else
-            {
-                /* COnstruct the user params recor for the PDF/x conversion */
-                PDFProcessorPDFXConvertParamsRec userParams;
-
-                memset ((char *)&userParams, 0, sizeof (PDFProcessorPDFXConvertParamsRec));
-                userParams.size = sizeof (PDFProcessorPDFXConvertParamsRec);
-                userParams.abortIfXFAPresent = false;
-                userParams.colorCompression = kPDFProcessorColorJpegCompression;
-                userParams.grayCompression = kPDFProcessorGrayZipCompression;
-                userParams.monoCompression = kPDFProcessorMonoCCITTGroup4Compression;
-                removeAllAnnotations[sequence % removeAllAnnotationsCount];
-
-            /* Create the ouput file ASPath name */
-#if !MAC_ENV	
-                ASPathName destFilePath = ASFileSysCreatePathName (NULL, ASAtomFromString ("Cstring"), fullOutputFileName, NULL);
-#else
-                ASPathName destFilePath = APDFLDoc::makePath (fullOutputFileName);
-#endif
-
-                /* Release the output file name */
-                free (fullOutputFileName);
-
-                /* Perform the conversions */
-                PDFProcessorConvertAndSaveToPDFX (inDoc.getPDDoc (), destFilePath, ASGetDefaultFileSys (), 
-                                                  ConvertOptions[convertorOptions[sequence % convertorOptionsCount]], &userParams);
-
-                /* Release the output path name */
-                ASFileSysReleasePath (NULL, destFilePath);
-
-                /* terminate the plugin */
-                PDFProcessorTerminate ();
-            }
-        HANDLER
-            info->result = 2;
-        END_HANDLER
-
-
-        if (!silent)
-            fprintf (info->logFile, "PDF/x Worker Thread Completed! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-
-    }
-
-private:
-
-    bool removeAllAnnotations[100];
-    int  removeAllAnnotationsCount;
-
-    ASUns32 convertorOptions[100];
-    int  convertorOptionsCount;
-
-    PDFProcessorPDFXConversionOption ConvertOptions[2];
-};
-
-#if !MAC_ENV    //yluo XPS2PDF plugin is not available for Mac platform
-/*
-** XPS2PDF worker
-** This thread will convert one document from XPS to PDF.
-** NoAPDFL MUST be false, or the thread will not function.
-** InFileName as a list will use the Nth entry of the list, modulus the number of entries in the list, for each worker thread.
-
-**  XPS2PDFOptions=[
-**                   silent=true                                        When true, do not display status lines
-**                   InFileName=[test data path\XPStoPDF.xps]      A list of file to copy, One file will be copied by each thread, sequcence % #files.
-**                   OutFilePath=[Output]                               Directory where output is written
-**                   noAPDFL=false                                      When true, we will NOT init/term the library for each thread
-*/
-#include "XPS2PDFCalls.h"
-class XPS2PDFWorker : public workerclass
-{
-public:
-    XPS2PDFWorker () { };
-    ~XPS2PDFWorker () { };
-
-    /* Parse the PDFtoXPS conversion thread options into attributes */
-    void ParseOptions (valuelist *values)
-    {
-        workerclass::ParseOptions(values, "%XPStoPDF.xps", "Output");
-
-        if (threadAttributes->IsKeyPresent ("LoadPlugins"))
-            workers[XPS2PDF].LoadPlugins = threadAttributes->GetKeyValueBool ("LoadPlugins");
-    };
-
-
-    /* One thread worker procedure */
-    void WorkerThread (ThreadInfo *info)
-    {
-        int sequence = info->sequence;
-        if (!silent)
-            fprintf (info->logFile, "XPS to PDF Worker Thread! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-
-        DURING
-
-            /* Set up the HFT for XPS2PDF */
-            gXPS2PDFHFT = InitXPS2PDFHFT;
-
-                /* Load and initialize the XPS2PDF plugin */
-            if (!XPS2PDFInitialize ())
-                info->result = 1;
-            else
-            {
-                /* Set up the job options. */
-                ASCab settings = ASCabNew ();
-
-                //The .joboptions file specifies a great number of settings which determine exactly how the PDF document
-                //is created by the converter.
-                ASText jobNameText = ASTextFromUnicode ((ASUTF16Val*)"../Resources/joboptions/Standard.joboptions", kUTF8);
-                ASCabPutText (settings, "PDFSettings", jobNameText);
-
-                //Specify which description in the .joboptions file we will use.
-                //There are many others, for different langauges. See the file.
-                ASText language = ASTextFromUnicode ((ASUTF16Val*)"ENU", kUTF8);
-                ASCabPutText (settings, "PDFSettingsLang", language);
-
-                /* Generate input and output file names */
-                char *fullFileName = GetInFileName (sequence);
-                char *fullOutputFileName = GetOutFileName (sequence, -1);
-
-                /* The automatic logic will use he same suffix for the output as the input, so change the suffix here */
-                char *suffix = &fullOutputFileName[strlen (fullOutputFileName) - 3];
-                suffix[0] = 0;
-                strcat (fullOutputFileName, "pdf");
-
-                /* Create the ASPathName for the input (XPS) document */
-                ASPathName asInPathName = ASFileSysCreatePathName (NULL, ASAtomFromString ("Cstring"), fullFileName, 0);
-
-                /* release the input file name */
-                free (fullFileName);
-
-                /* Convert the document */
-                PDDoc outputDoc = NULL;
-                int ret_val = XPS2PDFConvert (settings, 0, asInPathName, NULL, &outputDoc, NULL);
-
-                //If we succeeded, XPS2PDFConvert returns 1.
-                if (ret_val != 1)
-                    info->result = 2;
-                else
-                {
-                    /* Save the output PDF Document */
-                    APDFLDoc outAPDoc;
-                    outAPDoc.pdDoc = outputDoc;
-                    outAPDoc.saveDoc (fullOutputFileName);
-                }
-
-                /* release the output file name */
-                free (fullOutputFileName);
-
-                /* Release other resources created */
-                ASCabDestroy (settings);
-                ASFileSysReleasePath (NULL, asInPathName);
-
-            }
-
-            /* Terminate the plugin */
-            XPS2PDFTerminate ();
-
-        HANDLER
-                info->result = 3;
-        END_HANDLER
-
-        if (!silent)
-            fprintf (info->logFile, "XPS to PDF Worker Completed! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-    }
-
-
-};
-#endif
-
-
-/*
-** TextExtract  worker
-** This thread extract the content of N pages (Specified in page count, below)
-** NoAPDFL MUST be false, or the thread will not function.
-** InFileName as a list will use the Nth entry of the list, modulus the number of entries in the list, for each worker thread.
-
-**  PDFxOptions=[
-**                   silent=true                                        When true, do not display status lines
-**                   InFileName=[test data path\constitution.pdf]      A list of file to copy, One file will be copied by each thread, sequcence % #files.
-**                   OutFilePath=[Output]                               Directory where output is written
-**                   noAPDFL=false                                      When true, we will NOT init/term the library for each thread
-**                   NumberOfPages=[1]                                  Number of pages of content to extract in this thread (100 values max!)
-*/
-class TextextWorker : public workerclass
-{
-public:
-    TextextWorker () { };
-    ~TextextWorker () { };
-
-    /* Parse the text extract conversion thread options into attributes */
-    void ParseOptions (valuelist *values)
-    {
-        workerclass::ParseOptions (values, "%constitution.pdf", "Output");
-
-        if (threadAttributes->IsKeyPresent ("LoadPlugins"))
-            workers[TextExtract].LoadPlugins = threadAttributes->GetKeyValueBool ("LoadPlugins");
-
-        if (threadAttributes->IsKeyPresent ("NumberOfPages"))
-        {
-            valuelist *values = threadAttributes->GetKeyValue ("NumberOfPages");
-            pagesCount = values->size ();
-            for (ASUns32 index = 0; index < pagesCount; index++)
-                pages[index] = values->GetValueInt (index);
-        }
-        else
-        {
-            pagesCount = 1;
-            pages[0] = 1;
-        }
-
-    };
-
-
-    /* One thread worker procedure */
-    void WorkerThread (ThreadInfo *info)
-    {
-        int sequence = info->sequence;
-        if (!silent)
-            fprintf (info->logFile, "Text Extraction Worker Thread Started! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-
-        /* Generate input name */
-        char *fullFileName = GetInFileName (sequence);
-
-
-        DURING
-            /* Open the input document */
-            APDFLDoc inDoc (fullFileName, true);
-            PDDoc pdDoc = inDoc.getPDDoc ();
-
-
-            /* Get the number of pages */
-            size_t pagesInDocument = PDDocGetNumPages (pdDoc);
-
-            /* */
-            PDWordFinderConfigRec wfConfig;
-            memset (&wfConfig, 0, sizeof (PDWordFinderConfigRec));
-            wfConfig.recSize = sizeof (PDWordFinderConfigRec);
-
-            /* Create a word finder */
-            PDWordFinder wordFinder = PDDocCreateWordFinderEx (pdDoc, WF_LATEST_VERSION, false, &wfConfig);
-
-            /* How many pages have we already done*/
-            ASUns64 numberOfPagesDone = 0;
-            for (int index = 0; index < sequence; index++)
-                numberOfPagesDone += pages[index % pagesCount];
-
-            /* reduce modulos the number of pages in the document */
-            ASUns32 firstPageToDo = numberOfPagesDone % pagesInDocument;
-
-            ASUns32 numberOfPagesToDo = pages[sequence % pagesCount];
-
-            for (ASUns32 indexPage = 0; indexPage < pages[sequence % pagesCount]; indexPage++)
-            { 
-
-                /* Acquire the words for the Nth page */
-                PDWord wordList;
-                ASInt32 numWordsFound;
-                ASUns32 pageToDo = (firstPageToDo + indexPage) % pagesInDocument;
-
-                PDWordFinderAcquireWordList (wordFinder, pageToDo, &wordList, NULL, NULL, &numWordsFound);
-
-                /* The automatic logic will use he same suffix for the output as the input, so change the suffix here */
-                char *fullOutputFileName = GetOutFileName (sequence, pageToDo + 1);
-                char *suffix = &fullOutputFileName[strlen (fullOutputFileName) - 3];
-                suffix[0] = 0;
-                strcat (fullOutputFileName, "txt");
-
-                /* Create a file to hold the words */
-                FILE *wordFile = fopen (fullOutputFileName, "w");
-
-                /* Release the file name */
-                free (fullOutputFileName);
-
-                /* Write document and page name to result file */
-                fprintf (wordFile, "%s Page: %01d", fullFileName, pageToDo);
-
-
-                for (int i = 0; i < numWordsFound; ++i)
-                {
-                    /* Get the next word */
-                    PDWord nextWord = PDWordFinderGetNthWord (wordFinder, i);
-
-                    char workWord[1024];
-                    PDWordGetString (nextWord, workWord, 1024);
-                    fprintf (wordFile, "(%3d)  %s\n", i, workWord);
-                }
-
-                /* Close the word file */
-                fclose (wordFile);
-
-                  /* Release the word finder results */
-                PDWordFinderReleaseWordList (wordFinder, 0);
-            }
-
-            /* Release the word finder itself*/
-                PDWordFinderDestroy (wordFinder);
-
-            /* free input file name  */
-            free (fullFileName);
-
-        HANDLER
-            info->result = 1;
-        END_HANDLER
-
-        if (!silent)
-            fprintf (info->logFile, "Text Extraction Worker Thread Completed! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-    }
-
-    private:
-        ASUns32 pages[100];
-        ASUns32 pagesCount;
-
-};
-
-
-
-/*
-** Rasterizer Worker
-** This worker type will open a document, rasterize one or more pages, one or more times, Optoiopnally placiong the 
-** results into a new PDF file. 
-**
-** If "NoAPDFL=true" must not be specified for this worker type!
-**
-**  RasterizerOptions=[
-**                   silent=true                                        When true, do not display status lines
-**                   InFileName=[test data path\constitution.pdf]      A list of file to copy, One file will be copied by each thread, sequcence % #files.
-**                   OutFilePath=[Output]                               Directory where output is written
-**                   NumberOfPages=[1]                                  How many pages shall we rasterize (100 values max!)
-**                   Repetitions=[1]                                    How many times shall we rasterize each page done (100 values max!)
-**                   SaveImages=false                                   When true, we will save the inages in a PDf document, when false, we will not
-**                   Resolution=300                                     Resolution to render image too.
-**                   ColorModel={RGB,CMYK,GRAY,DeviceN]                 Which color model to use. RGB is the default.
-*/
-
-class RasterizerWorker : public workerclass
-{
-    /* This thread will copy a specified file N times, into new specified files.
-    ** Each time the file is copied, it will also find the fist 1000 prime numbers!
-    */
-public:
-    RasterizerWorker () { };
-    ~RasterizerWorker () { };
-
-    /* Parse the non-APDFL conversion thread options into attributes */
-    void ParseOptions (valuelist *values)
-    {
-        workerclass::ParseOptions (values, "%AddRedaction.pdf", "Output");
-
-        if (threadAttributes->IsKeyPresent ("LoadPlugins"))
-            workers[Rasterizer].LoadPlugins = threadAttributes->GetKeyValueBool ("LoadPlugins");
-
-        if (threadAttributes->IsKeyPresent ("Repetitions"))
-        {
-            valuelist *values = threadAttributes->GetKeyValue ("Repetitions");
-            RepetitionsCount = values->size ();
-            for (int index = 0; index < RepetitionsCount; index++)
-                Repetitions[index] = values->GetValueInt (index);
-        }
-        else
-        {
-            RepetitionsCount = 1;
-            Repetitions[0] = 5;
-        }
-
-        if (threadAttributes->IsKeyPresent ("NumberOfPages"))
-        {
-            valuelist *values = threadAttributes->GetKeyValue ("NumberOfPages");
-            pagesCount = values->size ();
-            for (ASInt32 index = 0; index < pagesCount; index++)
-                pages[index] = values->GetValueInt (index);
-        }
-        else
-        {
-            pagesCount = 1;
-            pages[0] = 1;
-        }
-
-        if (threadAttributes->IsKeyPresent ("SaveImages"))
-            saveImages = threadAttributes->GetKeyValueBool ("SaveImages");
-        else
-            saveImages = false;
-
-        if (threadAttributes->IsKeyPresent ("Resolution"))
-            Resolution = threadAttributes->GetKeyValueDouble ("Resolution");
-        else
-            Resolution = 300.0;
-
-        if (threadAttributes->IsKeyPresent ("ColorModel"))
-        {
-            valuelist *values = threadAttributes->GetKeyValue ("ColorModel");
-            char *name = values->value (0);
-            for (int i = 0;name[i] != 0; i++)
-                name[i] = toupper (name[i]);
-            if (!strcmp (name, "DEVICEN"))
-            {
-                strcpy (colorModel, name);
-                colorComponents = 4;
-            }
-            else
-            {
-                if (!strcmp (name, "DEVICEGRAY"))
-                {
-                    strcpy (colorModel, name);
-                    colorComponents = 1;
-                }
-                else
-                {
-                    if (!strcmp (name, "DEVICECMYK"))
-                    {
-                        strcpy (colorModel, name);
-                        colorComponents = 4;
-                    }
-                    else
-                    {
-                        strcpy (colorModel, "DeviceRGB");
-                        colorComponents = 3;
-                    }
-                }
-            }
-        }
-        else
-        {
-            strcpy (colorModel, "DeviceRGB");
-            colorComponents = 3;
-        }
-
-    };
-
-    char *RenderPageToBitmap (PDPage page, ASSize_t *mapSize, ASSize_t *width, ASSize_t *depth)
-    {
-
-        /* Get the matrix that transforms user space coordinates to rotated and cropped upright image coordinates.
-        **  The origin of this space is the top-left of the rotated, cropped page. Y is decreasing
-        */
-        ASFixedMatrix matrix;
-        PDPageGetFlippedMatrix (page, &matrix);
-
-        //Gets the crop box for a page. The crop box is the region of the page to display and print
-        ASFixedRect pageRect;
-        PDPageGetCropBox (page, &pageRect);
-
-        /* Set the scale matrix that will be concatenated to the user space matrix
-        ** to accomplish the desired resolution 
-        */
-        ASFixedMatrix scaleMatrix = { FloatToASFixed (Resolution / 72.0), 0, 0, FloatToASFixed (Resolution / 72.0), 0, 0 };
-
-        //Apply the scale to the default matrix
-        ASFixedMatrixConcat (&matrix, &scaleMatrix, &matrix);
-
-        /* Apply scale to the destination rectangle*/
-        ASFixedRect destRect;
-        ASFixedMatrixTransformRect (&destRect, &scaleMatrix, &pageRect);
-
-        /* If the page is offset, remove the offset here */
-        if (destRect.left != 0)
-        {
-            destRect.right -= destRect.left;
-            destRect.left = 0;
-        }
-        if (destRect.bottom != 0)
-        {
-            destRect.top -= destRect.bottom;
-            destRect.bottom = 0;
-        }
-
-        /* Generate width and depth */
-
-        /* Construct the draw params record */
-        PDPageDrawMParamsRec drawParams;
-        memset ((char *)&drawParams, 0, sizeof (PDPageDrawMParamsRec));
-        drawParams.size = sizeof (PDPageDrawMParamsRec);
-        drawParams.destRect = &destRect;
-        drawParams.matrix = &matrix;
-        drawParams.csAtom = ASAtomFromString(colorModel);
-        drawParams.bpc = 8;
-        drawParams.flags = kPDPageDoLazyErase;
-
-        /* Call draw to memory to get buffer size */
-        size_t bufferSize = PDPageDrawContentsToMemoryWithParams (page, &drawParams);
-
-        /* If we are doing deviceN, pickup the number of inks here */
-        if (drawParams.deviceNColorCount)
-            colorComponents = drawParams.deviceNColorCount[0];
-
-        /* If bufferSize is zero, then we set up inks inthe previous call (DeviceN)
-        ** So call again to get buffer size
-        */
-        if (bufferSize == 0)
-            bufferSize = PDPageDrawContentsToMemoryWithParams (page, &drawParams);
-
-        /* if buffer size is still zero, something is wrong. Error out!
-        */
-        if (bufferSize == 0)
-            return (NULL);
-
-        /* Allocate a buffer to hold the bitmap */
-        char *buffer = (char *)malloc (bufferSize);
-
-        /* Point to it int he draw params.*/
-        drawParams.buffer = buffer;
-        drawParams.bufferSize = bufferSize;
-
-        /* Draw the page !*/
-        PDPageDrawContentsToMemoryWithParams (page, &drawParams);
-
-        /* Tell the caller the size of the map, width, and depth */
-        *mapSize = bufferSize;
-        *width = abs (ASFixedRoundToInt16 (destRect.right) - ASFixedRoundToInt16 (destRect.left));
-        *depth = abs (ASFixedRoundToInt16 (destRect.top) - ASFixedRoundToInt16 (destRect.bottom));
-
-        /* Return the map to the caller */
-        return (buffer);
-    }
-
-    void AddImageToDoc (PDDoc doc, size_t mapSize, char *map, size_t width, size_t depth, ThreadInfo *info)
-    { 
-        /* Set upimage Attributes.
-        ** Always an XObject image
-        */
-        PDEImageAttrs attrs;
-        memset ((char *)&attrs, 0, sizeof (PDEImageAttrs));
-        attrs.flags = kPDEImageExternal;
-        attrs.width = width;
-        attrs.height = depth;
-        attrs.bitsPerComponent = 8;
-
-        /* Always compress image flate */
-        PDEFilterArray filterArray;
-        memset (&filterArray, 0, sizeof (PDEFilterArray));
-        filterArray.numFilters = 1;
-        filterArray.spec[0].name = ASAtomFromString ("FlateDecode");
-
-        /* Image is created passed to a 32 bit boundary per row.
-        ** For image usage, it must be padded to 8 bits per row
-        */
-        ASUns32 rowWidthPacked = width * colorComponents;
-        ASUns32 rowWidthPadded = (((width * colorComponents * 8) + 31) / 32) * 4;
-
-        /* Repack rows, as needed */
-        if (rowWidthPacked != rowWidthPadded)
-        {
-            /* Remove the row padding bytes */
-            for (ASUns16 row = 1; row < depth; row++)
-                memmove (&map[row * rowWidthPacked], &map[row * rowWidthPadded], rowWidthPacked);
-        }
-
-        /* New size */
-        size_t length = width * colorComponents * depth;
-
-        /* Image is erect, and sized 1 point per pixel */
-        ASFixedMatrix matrix = { width * fixedOne, 0, 0, depth * fixedOne, 0, 0 };
-
-        /* Color space as per creation
-        ** (NOTE: This will not work for deviceN color spaces, will have to 
-        **  add logic to create a color space from ink table is we want to support
-        ** deviceN colors. Too much to do for now.
-        */
-        PDEColorSpace cs = PDEColorSpaceCreateFromName (ASAtomFromString(colorModel));
-
-        /* Create a PDE Image (One that can be added to a PDF page */
-        PDEImage image = PDEImageCreate (&attrs, sizeof (attrs), &matrix, 0, cs, NULL, &filterArray, NULL, (ASUns8*)map, length);
-
-        /* Create a page to hold the image 
-        ** Just the size of the image.
-        */
-        ASFixedRect pageSize = { 0, depth * fixedOne, width * fixedOne, 0 };
-        PDPage page = PDDocCreatePage (doc, PDDocGetNumPages (doc)-1, pageSize);
-
-        /* Get the PDE Content */
-        PDEContent content = PDPageAcquirePDEContent (page, 0);
-
-        /* Set the image into the page */
-        PDEContentAddElem (content, kPDEAfterLast, (PDEElement)image);
-
-        /* Bind he content back into the page */
-        PDPageSetPDEContent (page, 0);
-
-        /* Release resources */
-        PDERelease ((PDEObject)image);
-        PDPageReleasePDEContent (page, 0);
-        PDERelease ((PDEObject)cs);
-
-    }
-
-    void WorkerThread (ThreadInfo *info)
-    {
-        int sequence = info->sequence;
-        if (!silent)
-            fprintf (info->logFile, "Rasterizer Worker Thread Started! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-
-        /* Generate input name */
-        char *fullFileName = GetInFileName (sequence);
-
-
-        DURING
-            /* Open the input document */
-            APDFLDoc inDoc (fullFileName, true);
-            PDDoc pdDoc = inDoc.getPDDoc ();
-
-
-            /* Get the number of pages */
-            size_t pagesInDocument = PDDocGetNumPages (pdDoc);
-
-            /* How many pages have we already done*/
-            ASUns64 numberOfPagesDone = 0;
-            for (int index = 0; index < sequence; index++)
-                numberOfPagesDone += pages[index % pagesCount];
-
-            /* reduce modulos the number of pages in the document */
-            ASUns32 firstPageToDo = numberOfPagesDone % pagesInDocument;
-
-            ASUns32 numberOfPagesToDo = pages[sequence % pagesCount];
-
-            PDDoc outDoc;
-            if (saveImages)
-                outDoc = PDDocCreate ();
-            else
-                outDoc = NULL;
-
-            /* Do the pages fromfirst to last, N times */
-            for (ASInt32 loop = 0; loop < Repetitions[sequence % RepetitionsCount]; loop++)
-            {
-                /* Do the requires set of pages */
-                for (ASInt32 indexPage = 0; indexPage < pages[sequence % pagesCount]; indexPage++)
-                {
-                    /* Obtain the current page */
-                    ASUns32 pageToDo = (firstPageToDo + indexPage) % pagesInDocument;
-                    PDPage page = PDDocAcquirePage (inDoc.getPDDoc(), pageToDo);
-
-                    /* Render the current page */
-                    ASSize_t mapsize, width, depth;
-                    char *mapBuffer = RenderPageToBitmap (page, &mapsize, &width, &depth);
-
-                    /* Release the current page */
-                    PDPageRelease (page);
-
-                    /* If we are saving the images, make the map and image, and 
-                    ** add it to the images document
-                    */
-                    if (saveImages)
-                    {
-                        AddImageToDoc (outDoc, mapsize, mapBuffer, width, depth, info);
-                    }
-
-                    /* Free the bitmap */
-                    free (mapBuffer);
-
-                }
-            }
-
-            if (saveImages)
-            {
-                /* The automatic logic will use he same suffix for the output as the input, so change the suffix here */
-                char *fullOutputFileName = GetOutFileName (sequence, -1);
-#if !MAC_ENV	
-                ASPathName destFilePath = ASFileSysCreatePathName (NULL, ASAtomFromString ("Cstring"), fullOutputFileName, NULL);
-#else
-                ASPathName destFilePath = APDFLDoc::makePath (fullOutputFileName);
-#endif
-
-                PDDocSave (outDoc, PDSaveFull | PDSaveCollectGarbage, destFilePath, NULL, NULL, NULL);
-                ASFileSysReleasePath (NULL, destFilePath);
-                free (fullOutputFileName);
-                PDDocClose (outDoc);
-            }
-
-        HANDLER
-            info->result = 1;
-        END_HANDLER
-
-        if (!silent)
-            fprintf (info->logFile, "Rasterizer Worker Thread completed! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-    }
-
-private:
-    ASInt32     Repetitions[100];
-    ASInt32     RepetitionsCount;
-    ASInt32     pages[100];
-    ASInt32     pagesCount;
-    ASBool      saveImages;
-    double      Resolution;
-    char        colorModel[20];
-    ASInt8      colorComponents;
-};
-
-/*
-** Flattener worker
-** This thread will flatten document.
-** NoAPDFL MUST be false, or the thread will not function.
-** InFileName as a list will use the Nth entry of the list, modulus the number of entries in the list, for each worker thread.
-
-**  FlattenOptions=[
-**                   silent=true                                        When true, do not display status lines
-**                   InFileName=[test data path\AddRedactions.pdf]      A list of file to copy, One file will be copied by each thread, sequcence % #files.
-**                   OutFilePath=[Output]                               Directory where output is written
-**                   noAPDFL=false                                      When true, we will NOT init/term the library for each thread
-*/
-#include "PDFlattenerCalls.h"
-#include <iomanip>
-ASBool flattenerProgMon(ASInt32 pageNum, ASInt32 totalPages, float current, ASInt32 reserved, void *clientData)
-{
-    //The previous page we were working on.
-    ASInt32* prevPage = (ASInt32*)clientData;
-
-    //If we've begun a new page, or if we've finished.
-    if (pageNum != (*prevPage) || current == 100.0f)
-    {
-        //Print the completion percentage.
-        std::cout << "[" << std::fixed << std::setw(6) << std::setfill('0') << std::setprecision(2) << current << "%] ";
-
-        //Print the current page.
-        std::cout << "Flattening page " << pageNum + 1 << " of " << totalPages << ". " << std::endl;
-
-        //Update previous page.
-        *prevPage = pageNum;
-    }
-
-    //Return 1 to cancel Flattening.
-    return 0;
-}
-
-class FlattenWorker : public workerclass
-{
-#define NumberOfPDFAConvertOptions 4
-
-public:
-    FlattenWorker() {
-    };
-
-    ~FlattenWorker()
-    { };
-
-    /* Parse the PDF/a conversion thread options into attributes */
-    void ParseOptions(valuelist *values)
-    {
-        workerclass::ParseOptions(values, "%AddRedaction.pdf", "Output");
-
-        if (threadAttributes->IsKeyPresent ("LoadPlugins"))
-            workers[Flattener].LoadPlugins = threadAttributes->GetKeyValueBool ("LoadPlugins");
-    };
-
-    /* One thread worker procedure */
-    void WorkerThread(ThreadInfo *info)
-    {
-        int sequence = info->sequence;
-        if (!silent)
-            fprintf(info->logFile, "PDF/a Worker Thread Started! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-
-        /* Generate input and output file names */
-        char *fullFileName = GetInFileName(sequence);
-        char *fullOutputFileName = GetOutFileName(sequence, -1);
-
-        DURING
-        /* Open the input document */
-        APDFLDoc inDoc(fullFileName, true);
-
-        /* Free the input file names */
-        free(fullFileName);
-
-        /* Initialize the HFT for the plugin */
-        gPDFlattenerHFT = InitPDFlattenerHFT;;
-
-        //initialize PDFProcessor plugin
-        if (!PDFlattenerInitialize())
-            /* If the plugin cannot be initilized! */
-            info->result = 1;
-        else
-        {
-            /* COnstruct the user params recor for the flattening */
-            PDFlattenerUserParamsRec flattenParams;
-
-            memset(&flattenParams, 0, sizeof(PDFlattenerUserParamsRec));
-            flattenParams.size = sizeof(PDFlattenerUserParamsRec);
-            //A profiled color space to use for transparent objects. For CMYK, use "U.S. Web Coated (SWOP)v2".
-            flattenParams.profileDesc = ASTextFromUnicode((ASUTF16Val*)"sRGB IEC61966-2.1", kUTF8);
-            //The ZIP compression scheme (Flate encoding) for images.
-            flattenParams.colorCompression = kPDFlattenerZipCompression;
-            //Raster/Vector balance. Use 0.00f for no vectors.
-            flattenParams.transQuality = 100.0f;
-            // Callback options.
-
-            ASInt32 currentPage = -1;
-            //Progress monitor callback data. I'm using this data to store the previous page 
-            //   the Flattener was working on, using -1 as "hasn't begun yet".
-            flattenParams.progressClientData = (void*)&currentPage;
-            //The progress monitor callback function.
-            flattenParams.flattenProgress = flattenerProgMon;
-            // Tile flattening options
-
-            PDFlattenRec flattener;
-            memset(&flattener, 0, sizeof(PDFlattenRec));
-            flattener.size = sizeof(PDFlattenRec);
-
-            flattener.tilingMode = kPDNoTiling;
-            flattener.tileSizePts = 0;
-
-            //Resolution for flattening the interior of an atomic region.
-            flattener.internalDPI = 800.0f;
-            //Resolution for flattening edges of atomic regions.
-            flattener.externalDPI = 200.0f;
-
-            flattener.clipComplexRegions = false;
-            //If we convert stroked elements to filled elements.
-            flattener.strokeToFill = true;
-            //If we use rastered text instead of native text.
-            flattener.useTextOutlines = false;
-            //If we attempt to preserve overprint
-            flattener.preserveOverprint = true;
-
-            flattener.allowShadingOutput = true;
-            flattener.allowLevel3ShadingOutput = true;
-
-            //Maximum image size while flattening. 0 is default.
-            flattener.maxFltnrImageSize = 0;
-            //Adaptive flattening threshold. Doesn't matter, since we're not doing adaptive tiling. See tilingMode.
-            flattener.adaptiveThreshold = 0;
-
-            flattenParams.flattenParams = &flattener;
-
-            /* Perform the flatten document */
-            ASUns32 numFlattened = 0;
-            ASInt32 flattenResult = PDFlattenerConvertEx2(inDoc.getPDDoc(),      //The document whose pages we wish to flatten.
-                                                          0,                     //The first page to flatten.
-                                                          PDDocGetNumPages(inDoc.getPDDoc()) - 1,   //The last page to flatten.
-                                                          &numFlattened,         //PDFlattener sets this to the number of pages
-                                                          // it flattened. It will not flatten pages that do not contain transparent elements.
-                                                          &flattenParams);             //Flattener options.
-            if (flattenResult)
-            {
-                std::cout << "Flattened " << numFlattened << " pages." << std::endl;
-            }
-            else
-            {
-                std::cout << "Flattening failed." << std::endl;
-                ASRaise(GenError(genErrGeneral));
-            }
-            printf("outputfile name: %s\n", fullOutputFileName);
-            /* Save the output PDF Document */
-            inDoc.saveDoc(fullOutputFileName); //fullOutputFileName
-            /* Release the output file name */
-            free(fullOutputFileName);
-            ASTextDestroy(flattenParams.profileDesc);
-            /* terminate the plugin */
-            PDFlattenerTerminate();
-        }
-        HANDLER
-            info->result = 2;
-        END_HANDLER
-
-            if (!silent)
-                fprintf(info->logFile, "PDF/a Worker Thread Completed! (Sequence: %01d, Thread: %01d\n", sequence + 1, info->threadNumber + 1);
-    }
-
-};
-
-
 /* This procedure calls a worker thread of a specific type, 
-** based on the info->objectType field. 
+** based on the GetWorkerClass method. 
 **
 ** This must be updated for every new worker type!
 */
@@ -1504,7 +158,7 @@ void callWorker (ThreadInfo *info)
 {
     workerclass *baseObject = (workerclass *)(info->object);
 
-    switch (info->objectType)
+    switch (baseObject->GetWorkerClass())
     {
     case NONAPDFL:
         ((NonAPDFLWorker *)baseObject)->WorkerThread (info);
@@ -1515,11 +169,9 @@ void callWorker (ThreadInfo *info)
     case PDFX:
         ((PDFxWorker *)baseObject)->WorkerThread (info);
         break;
-#if !MAC_ENV
     case XPS2PDF:
         ((XPS2PDFWorker *)baseObject)->WorkerThread (info);
         break;
-#endif
     case TextExtract:
         ((TextextWorker *)baseObject)->WorkerThread (info);
         break;
@@ -1668,26 +320,25 @@ int main(int argc, char** argv)
     /* Construct the array of worker types 
     ** Set establish the options for each type*/
     workerClasses[NONAPDFL].NonAPDFL = new NonAPDFLWorker ();
-    workerClasses[NONAPDFL].NonAPDFL->ParseOptions (SampleAttributes.GetKeyValue (workers[NONAPDFL].paramName));
-
+    workerClasses[NONAPDFL].NonAPDFL->ParseOptions (&SampleAttributes, &workers[NONAPDFL]);
 
     workerClasses[PDFA].PDFa = new PDFaWorker ();
-    workerClasses[PDFA].PDFa->ParseOptions (SampleAttributes.GetKeyValue (workers[PDFA].paramName));
+    workerClasses[PDFA].PDFa->ParseOptions (&SampleAttributes, &workers[PDFA]);
 
     workerClasses[PDFX].PDFx = new PDFxWorker ();
-    workerClasses[PDFX].PDFx->ParseOptions (SampleAttributes.GetKeyValue (workers[PDFX].paramName));
-#if !MAC_ENV    //XPS2PDF does not supported by MAC
+    workerClasses[PDFX].PDFx->ParseOptions (&SampleAttributes, &workers[PDFX]);
+
     workerClasses[XPS2PDF].XPS2PDF = new XPS2PDFWorker ();
-    workerClasses[XPS2PDF].XPS2PDF->ParseOptions (SampleAttributes.GetKeyValue (workers[XPS2PDF].paramName));
-#endif
+    workerClasses[XPS2PDF].XPS2PDF->ParseOptions (&SampleAttributes, &workers[XPS2PDF]);
+
     workerClasses[TextExtract].TextExtract = new TextextWorker ();
-    workerClasses[TextExtract].TextExtract->ParseOptions (SampleAttributes.GetKeyValue (workers[TextExtract].paramName));
+    workerClasses[TextExtract].TextExtract->ParseOptions (&SampleAttributes, &workers[TextExtract]);
 
     workerClasses[Rasterizer].Rasterizer = new RasterizerWorker ();
-    workerClasses[Rasterizer].Rasterizer->ParseOptions (SampleAttributes.GetKeyValue (workers[Rasterizer].paramName));
+    workerClasses[Rasterizer].Rasterizer->ParseOptions (&SampleAttributes, &workers[Rasterizer]);
 
     workerClasses[Flattener].Flattener = new FlattenWorker();
-    workerClasses[Flattener].Flattener->ParseOptions(SampleAttributes.GetKeyValue(workers[Flattener].paramName));
+    workerClasses[Flattener].Flattener->ParseOptions (&SampleAttributes, &workers[Flattener]);
 
     /* This will be the list of threads to run */
     ThreadInfo *threads = (ThreadInfo *)malloc (sizeof (ThreadInfo) * totalThreads);
@@ -1747,7 +398,6 @@ int main(int argc, char** argv)
         threads[index].threadNumber = index;
         threads[index].sequence = index / processes;
         threads[index].object = (void *)workerList[type].PDFa;
-        threads[index].objectType = workerTypeList[type];
         threads[index].logFile = logFile;
         threads[index].logFileSet = logFileSet;
         threads[index].LoadPlugins = workers[workerTypeList[type]].LoadPlugins;
