+++ Multi-Threading/MultiThreadingSample/PDFx.txt
@@ -1,0 +1,15 @@
+TotalThreads=10
+activeThreads=5 
+BaseInit=true 
+processes=PDFx
+PauseEvery=0
+TempMemFileSys=false
+PDFxOptions=[
+       silent=false, 
+       noapdfl=false, 
+       InFileName=%trans_1page.pdf,
+       OutFilePath=Output,
+       ConvertOption=3,
+       LoadPlugins=true,
+       RemoveAllAnnotations=false
+                     ]
